requirements.txt
cached-property==1.5.2
certifi==2021.10.8
charset-normalizer==2.0.10
frozendict==2.1.3
genanki==0.11.0
idna==3.3
mysql-connector-python==8.0.27
protobuf==3.19.3
pystache==0.6.0
PyYAML==6.0
requests==2.27.1
urllib3==1.26.8
